Adding test to avoid duplicate scrapes: incomplete

+++ web_scraper/scraper.py
@@ -1,16 +1,16 @@
-# from nturl2path import url2pathname
 import requests
 from bs4 import BeautifulSoup
 from selenium import webdriver
 import urllib.request
 from webdriver_manager.chrome import ChromeDriverManager
+from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.common.by import By
 import os
 import uuid
 import json
 import boto3
 import pandas as pd
-import time
+# import time
 from sqlalchemy import create_engine
 from sqlalchemy import inspect
 from alive_progress import alive_bar
@@ -55,14 +55,20 @@ class Scraper:
             "chess_federation": []
         }
         # selenium webdriver for wiki scraping
-        self.driver = webdriver.Chrome(ChromeDriverManager().install())
+        options = Options()
+        options.add_argument('--headless')
+        options.add_argument('--disable-gpu')
+        self.driver = webdriver.Chrome(ChromeDriverManager().install(),
+                                       chrome_options=options)
         # set storage location
         self.data_store = "./raw_data"
+
+        # acquire existing data from RDS for comparison
+        ## TODO get RDS connection. Compare name and rank of existing before downloading new.
 
     def page_grab(self, url):
         self.url = url
         self.driver.get(self.url)
-        time.sleep(0.5)
         # requests and beautiful soup for getting the list of links
         page = requests.get(self.url)
         html = page.text
@@ -81,10 +87,13 @@ class Scraper:
         None
 
         """
+        print("Acquiring page of players")
         player_table = self.soup.find(name='tbody')
-        for link in player_table.find_all('a'):
-            self.player_data["links"].append(link.get('href'))
-            self.player_data["uuid"].append(str(uuid.uuid4()))
+        with alive_bar(50, bar='smooth') as bar:
+            for link in player_table.find_all('a'):
+                self.player_data["links"].append(link.get('href'))
+                self.player_data["uuid"].append(str(uuid.uuid4()))
+                bar()
 
     def create_store(self, folder):
         """Checks to see if the raw_data folder exists, if not, creates it.
@@ -179,51 +188,54 @@ class Scraper:
                      "chess_federation":
                      self.player_data["chess_federation"][index]}
                 ]
-                self.data_dump(folder_name, data)
+                # create the data.json from the above dictionary
+                with open(f"raw_data/{folder_name}/data.json", "w") as f:
+                    json.dump(data, f)
+                
+                self.data_dump(folder_name) #AWS S3 UPLOAD
                 bar()
 
     def search_wiki(self, index, folder_name):
         search_term = self.player_data["search_term"][index]
-        self.driver.get("http://wikipedia.org")
-        time.sleep(0.5)
-        # choose search box
-        search = self.driver.find_element_by_id("searchInput")
-        search.click()
-        time.sleep(0.5)
-        # magnus is famous enough to have his own Wiki URL
-        # so I need to fudge his search.
-        # if any other player gets their own web direct link
-        # it'll break this scraper.
-        if self.player_data["name"][index] == "Magnus Carlsen":
-            search.send_keys("Magnus Carlsen Norwegian")
-        else:
-            search.send_keys(search_term)
-        confirm = self.driver.find_element_by_xpath(
-            "/html/body/div[3]/form/fieldset/button")
-        confirm.click()
-        # try the top result
-        try:
-            first_result = self.driver.find_element_by_xpath(
-                "/html/body/div[3]/div[3]/div[4]/div[3]/ul/li[1]/div[1]/a")
-        except Exception:
+        if not os.path.isfile(f'raw_data/{folder_name}/{folder_name} - wiki.jpg'):
+            self.driver.get("http://wikipedia.org")
+            # choose search box
+            search = self.driver.find_element_by_id("searchInput")
+            search.click()
+            # magnus is famous enough to have his own Wiki URL
+            # so I need to fudge his search.
+            # if any other player gets their own web direct link
+            # it'll break this scraper.
+            if self.player_data["name"][index] == "Magnus Carlsen":
+                search.send_keys("Magnus Carlsen Norwegian")
+            else:
+                search.send_keys(search_term)
+            confirm = self.driver.find_element_by_xpath(
+                "/html/body/div[3]/form/fieldset/button")
+            confirm.click()
+            # try the top result
             try:
-                # some pages have a "did you mean" to ignore
                 first_result = self.driver.find_element_by_xpath(
-                    ("/html/body/div[3]/div[3]/div[4]/" +
-                        "div[4]/ul/li[1]/div[1]/a"))
+                    "/html/body/div[3]/div[3]/div[4]/div[3]/ul/li[1]/div[1]/a")
             except Exception:
+                try:
+                    # some pages have a "did you mean" to ignore
+                    first_result = self.driver.find_element_by_xpath(
+                        ("/html/body/div[3]/div[3]/div[4]/" +
+                            "div[4]/ul/li[1]/div[1]/a"))
+                except Exception:
+                    pass
+            try:
+                self.driver.find_element(By.CLASS_NAME, 'mw-search-nonefound')
                 pass
-        try:
-            self.driver.find_element(By.CLASS_NAME, 'mw-search-nonefound')
-            pass
-        except Exception:
-            link = first_result.get_attribute("href")
-            self.driver.get(link)
-            self.pull_image(folder_name, search_term)
-        # data_dump() and pull_image() both need to run on
-        # every iteration of a person search
+            except Exception:
+                link = first_result.get_attribute("href")
+                self.driver.get(link)
+                self.pull_image(folder_name)
+            # data_dump() and pull_image() both need to run on
+            # every iteration of a person search
 
-    def pull_image(self, folder_name, search_term):
+    def pull_image(self, folder_name):
         """
         Downloads the image from the specific location on the wikipedia page.
 
@@ -263,37 +275,30 @@ class Scraper:
             self.create_store(f'raw_data/{folder_name}')
             image = image.get_attribute("src")
             urllib.request.urlretrieve(
-                image, f"raw_data/{folder_name}/{folder_name}.jpg")
+                image, f"raw_data/{folder_name}/{folder_name} - wiki.jpg")
 
-    def data_dump(self, folder_name, data):
+    def data_dump(self, folder_name):
         """
-        Converts the player_data into a json file and stores it
 
         Parameters:
         ----------
         folder_name: str
             String value of the folder path for each player's data store
 
-        data: dict
-            All data for the given player, in a dictionary
-
         Returns:
         -------
         None
         """
-        with open(f"raw_data/{folder_name}/data.json", "w") as f:
-            json.dump(data, f)
-        # filename = (f"raw_data/{folder_name}/data.json")
-        # try:
-        #    pic_file = (f"raw_data/{folder_name}/{folder_name}.jpg")
-        # except Exception:
-        #    pic_file = ""
-        # try:
-        #    pic_file2 = (f"raw_data/{folder_name}/{folder_name}2.jpg")
-        # except Exception:
-        #    pic_file2 = ""
-        # self.upload_to_aws(
-        #    filename, 'chess-top-50', pic_file, pic_file2, folder_name)
+        try:
+            pic_file = (f"raw_data/{folder_name}/{folder_name}.jpg")
+        except Exception:
+            pic_file = ""
+        try:
+            pic_file2 = (f"raw_data/{folder_name}/{folder_name} - wiki.jpg")
+        except Exception:
+            pic_file2 = ""
+        self.upload_to_aws((f"raw_data/{folder_name}/data.json"),
+                           'chess-top-50', pic_file, pic_file2, folder_name)
 
     def follow_links_more_data(self, name, link):
         """Going to the individual page on chess.com and getting extra data
@@ -301,26 +306,27 @@ class Scraper:
         This downloads the player date of birth and a second image, as well
         as country of origin and chess federation that the player represents.
         """
-        self.driver.get(link)
-        player_table = self.driver.find_elements(
-            By.CLASS_NAME, "master-players-value")
-        player_stats = []
-        for row in player_table:
-            player_stats.append(row.text)
-        self.player_data["date_of_birth"].append(player_stats[1])
-        self.player_data["place_of_birth"].append(player_stats[2])
-        self.player_data["chess_federation"].append(player_stats[3])
-        # Get second image
-        image = self.driver.find_element(By.CLASS_NAME,
-                                         "post-view-thumbnail")
-        image = image.get_attribute("src")
-        image = ("https://www.chess.com/bundles/web/" +
-                 "images/user-image.007dad08.svg")
-        try:
-            urllib.request.urlretrieve(image, f"raw_data/{name}/{name}2.jpg")
-        except Exception:
-            urllib.request.urlretrieve("https://tinyurl.com/y7l4drrj",
-                                       f"raw_data/{name}/{name}2.jpg")
+        if not os.path.isfile(f'raw_data/{name}/data.json'):
+            self.driver.get(link)
+            player_table = self.driver.find_elements(
+                By.CLASS_NAME, "master-players-value")
+            player_stats = []
+            for row in player_table:
+                player_stats.append(row.text)
+            self.player_data["date_of_birth"].append(player_stats[1])
+            self.player_data["place_of_birth"].append(player_stats[2])
+            self.player_data["chess_federation"].append(player_stats[3])
+            # Get chess.com image if not already acquired
+            try:
+                image = self.driver.find_element(By.CLASS_NAME,
+                                                "post-view-thumbnail")
+                image = image.get_attribute("src")
+                urllib.request.urlretrieve(image, f"raw_data/{name}/{name}.jpg")
+            except Exception:
+                urllib.request.urlretrieve("https://tinyurl.com/y7l4drrj",
+                                        f"raw_data/{name}/{name}.jpg")
+        else:
+
 
     def upload_to_aws(self, filename, bucket, image, image2, folder):
         """Gets the data.json file and the image.jpg and uploads them to AWS"""
@@ -337,7 +343,7 @@ class Scraper:
                 pass
             try:
                 s3b.meta.client.upload_file(
-                    image2, bucket, f"raw_data/{folder}/{folder}2.jpg")
+                    image2, bucket, f"raw_data/{folder}/{folder} - wiki.jpg")
                 print("uploaded second picture")
             except Exception:
                 pass
@@ -401,8 +407,9 @@ if __name__ == "__main__":
             chess_scrape.create_store(chess_scrape.data_store)
             chess_scrape.get_player_data()
             # chess_scrape.driver.close()
-        chess_scrape.wiki_player_search()
         chess_scrape.upload_table_data()
+        chess_scrape.wiki_player_search()
+        
     else:
         chess_scrape = Scraper("http://chess.com/ratings")
         chess_scrape.upload_table_data()
